+++ src/structs.h
@@ -48,9 +48,9 @@
  * on an older version. You are supposed to compare this with the macro
  * LUMINARIMUD_VERSION() in utils.h.
  * It is read as Major/Minor/Patchlevel - MMmmPP (hex values)
- * Current: 0x030704 = 3.7.4 (corresponds to display version 2.5034-beta)
+ * Current: 0x030705 = 3.7.5 (corresponds to display version 2.5035-beta)
  * See also: src/constants.c (luminari_version string) */
-#define _LUMINARIMUD 0x030704
+#define _LUMINARIMUD 0x030705
 
 /** If you want equipment to be automatically equipped to the same place
  * it was when players rented, set the define below to 1 because